refactor(agent): extract runtime module (containerd + dockerd lifecycle)
Pull the entire runtime lifecycle out of linux/mod.rs into runtime.rs:

- EnsureRuntime / RuntimeStatus RPC handlers + the start/probe drivers.
- containerd / dockerd spawn + readiness polling
  (`ensure_containerd_ready`, `ensure_dockerd_ready`, `collect_runtime_status`).
- bundled-runtime driver (`try_start_bundled_runtime`) and its lock.
- helpers: `runtime_path_env`, `ensure_shared_runtime_dirs`,
  `shared_containerd_config`, `daemon_log_file`, the binary-presence
  detectors (`detect_runtime_bin_dir`, `runtime_missing_detail*`,
  `missing_runtime_binaries_at`, `missing_binaries_at`).
- prerequisite mounter (`ensure_runtime_prerequisites`) plus the
  `MIN_SANE_EPOCH` clock-guard constant and `CONTAINERD_SOCKET_CANDIDATES`.

Items shared with the upcoming kubernetes module are exposed `pub(super)`
(daemon_log_file, ensure_containerd_ready, ensure_runtime_prerequisites,
ensure_shared_runtime_dirs, runtime_path_env, missing_binaries_at,
CONTAINERD_SOCKET_CANDIDATES). `sync_clock_from_host` stays in mod.rs
for now (relocates with the rpc module in a later commit) and gains
`pub(super)` so runtime.rs can reach it.

The `test_shared_containerd_config_uses_k3s_cni_paths` test moves to
runtime.rs's own `#[cfg(test)] mod tests`, dropping the cross-module
private access that motivated the earlier `pub(super)` patch.

Pure mechanical extraction. Same warning set as before.
